Introduced Common.Client
