+++ src/TagzApp.Common.Client/Card.cs
