+++ src/TagzApp.Common.Client/ContentType.cs
