+++ src/TagzApp.Common.Client/ModerationState.cs
